Added data.ts which is essentially data.json except in a TypeScript file
Ran into issues importing JSON into TypeScript, was of type Module but not Array as I would have expected. Since this assignment is time limited I'm going to cheat a little bit to get to the core of the project.

I have faced this problem before and was able to figure it out in the past but I have forgotten how I did it.