Fixed a React error, added missing key to a line in the table
When we map over a list of object and return JSX React expects the elements to have a unique key. The data we have doesn't have any obvious unique ID that I can see so I fall back to the index (0, 1, 2, ..) where the element is in the array.

Removed the `any` type from line. The type is now correctly inferred so it was redundant.