Added highcharts as a third party dependency
I was originally going to go with recharts for doing graphs but it seems that it doesn't have as good of a built-in support for potentially disjointed timescales.

I wanted to try to solve the problem as generally as possible and assume the least I could (within time limits) about the shape of the data.